requirements.txt
async-timeout==4.0.2
bidict==0.22.0
click==8.0.4
Deprecated==1.2.13
dnspython==2.2.1
eventlet==0.33.0
Flask==2.0.3
Flask-SocketIO==5.1.1
Flask-WTF==1.0.0
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
packaging==21.3
pyparsing==3.0.7
python-engineio==4.3.1
python-socketio==5.5.2
six==1.16.0
typing_extensions==4.1.1
Werkzeug==2.0.3
wrapt==1.14.0
WTForms==3.0.1
beautifulsoup4
requests
